# btf: read all line info records at once in `parseLineInfoRecords`