@ti-mo 2025-01-27
Cool! Drive-by: whenever you optimize old code that doesn't change frequently, make sure to bolt it down with `testing.AllocsPerRun`. :wink: