@paulcacheux 2025-01-27
I added a small test for the allocations, I'm a bit sad about it because it checks against 3:
- as opposed to all other allocation tests in the repo that test against 0
- it feels a bit fragile, since it's susceptible to changes in the stdlib that would improve or degrade the amount of allocations of `binary.Read`
Otherwise I could do something like `< 10` or something like that, WDYT ?