@dylandreimerink 2025-01-27
> it feels a bit fragile, since it's susceptible to changes in the stdlib that would improve or degrade the amount of allocations of binary.Read

Yea, I would have to agree with that, seems fragile. I imagine it would be frustrating if we have to start updating magic values in out tests when we bump Go / stdlib versions. I don't know if a <10 version hits the mark, like if we go to 11, would that be a cause for us to reconsider something?

On the other hand, I understand the desire to curb allocations and memory use caused by the lib. Not sure if this is the best way to do so however.