@ti-mo 2025-01-29
Discussed this offline with Dylan. I agree on all arguments raised, but:

- we have a few of these tests in the code base, and so far they've been quiet
- let's assume that amount of allocs in stdlib functions shrinks over time instead of growing, given the amount of scrutiny on them
- these alloc measurements are enforced over small blocks of deterministic code without side effects
- we can always remove them if they cause too much noise (that would be on me :slightly_smiling_face:)

I think pinning it down to 7 allocs is the best we can do in absence of a continuous benchmarking setup or a solution to track allocs of certain spans of code over time. These alternatives are (much) more expensive to run and not maintenance-free by any means.

Also, since this is a wire format defined in a (sort of) spec that doesn't change on a whim, my tendency would be to write a custom marshaler for these things, but I've had trouble convincing Lorenz of this in the past. If we really need this to not allocate, it's still the way to go IMO.