@paulcacheux 2025-01-29
Ok, I can change the test to 7 for this PR. And work on a follow up PR with a real unmarshaller